Implement rounds - the game can now be played!
After every player has rolled the dice for a
category, it becomes the next player's turn to pick a
new category.

Also, simplify the broadcast of scores. We just
broadcast the updated scoreboard for every incoming
update.

Dice roll messages need the username so that we only
count a user's dice rolls once. Counting who has
rolled the dice for a round correctly is required in
order to implement rounds and turns correctly. If I
get two dice results from a single player, it does
not mean that two players have rolled the dice, it's
just because we have multiple websocket connections
from each client.