Store player name in cookie
This makes it possible to refresh the lobby and still
retain your identity in the lobby and the ensuing
game once it is started.

Also, use a different scheme for storing the host
marker cookie to make it more tied to the particular
game they are the host of.

Improve variable names in Lobby and generally do some
polishing of it.

Don't show the "Your name:" bit until a player has
entered their name, to avoid confusion.

Added a couple of missing tests for the lobby
regarding the state of the Start Game button.